first version with merge mode
with this, it's possible to add surfaces and zones to MEI files with existing content. This isn't ideal yet, though, as you can only add zones from one by one from the beginning of the file, and there are no options to change zones once they're added. This will require some revisions to the interaction of modes, but at least, this is a working proof of concept :)